fix: don't update based on parent offset on mount
* I'm not sure if this will work for all use cases, but for me it is workig much better, before this fix
there was an issue when the component position was stored externally and the component is remounted with a
defaultPos - it will jump by the parent offset pixels, which just isn't right. In my eyes it should just use
the exact position passed to it from consumers as the starting position, if a consumer wants to use the absolute
value they can pass that, otherwise the position they use `lastX` `lastY` for example is relative to the parent
anyway. So whats the point of applying this offset to those values?